% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_bern}
\alias{bin_bern}
\title{Create long data frame}
\usage{
bin_bern(df, Successes, Trials)
}
\arguments{
\item{Successes, Trials}{Numbers.}
}
\value{
"invisible" \code{NULL}.
}
\description{
This function takes data presented in Binomial (successes, trials) and converts to Bernoulli trials (0,1)
}
