% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cannabis.R
\docType{data}
\name{cannabis}
\alias{cannabis}
\title{High school cannabis use}
\format{
A tibble with 400 rows and 33 variables:
\describe{
\item{"failed_0"}{Academic grade (failed; 0: no failed subjects; failed at least one subject) at baseline, grade 7, wave 0}
\item{"failed_1"}{fail at baseline, grade 8, wave 1}
\item{"failed_2"}{fail at baseline, grade 9, wave 2}
\item{"failed_3"}{fail at baseline, grade 10, wave 3}
\item{"peer_can_0"}{peer cannabis use at grade 7 baseline (peer_can: 0: None; 1: 1-2 friends used cannabis; 2: 3 or more friends used cannabis)}
\item{"peer_can_1"}{peer cannabis use at grade 8}
\item{"peer_can_2"}{peer cannabis use at grade9}
\item{"peer_can_2"}{peer cannabis use at grade10}
\item{"peer_alc_0"}{peer alcohol use at grade 7 baseline}
\item{"peer_alc_1"}{peer alcohol use at grade 8}
\item{"peer_alc_2"}{peer alcohol use at grade 9}
\item{"peer_alc_3"}{peer alcohol use at grade 10}
\item{"antisocial_0"}{Antisocial behaviour (antisocial: 0: No antisocial behaviour; 1: engaged in antisocial behaviour) at wave 0}
\item{"antisocial_1"}{Antisocial behaviour (antisocial: 0: No antisocial behaviour; 1: engaged in antisocial behaviour) at wave 1}
\item{"antisocial_2"}{Antisocial behaviour (antisocial: 0: No antisocial behaviour; 1: engaged in antisocial behaviour) at wave 2}
\item{"antisocial_3"}{Antisocial behaviour (antisocial: 0: No antisocial behaviour; 1: engaged in antisocial behaviour) at wave 3}
\item{"alc_0"}{alcohol use at grade 7 baseline wave 0}
\item{"alc_1"}{alcohol use at grade 8 baseline wave 1}
\item{"alc_2"}{alcohol use at grade 9 baseline wave 2}
\item{"alc_3"}{alcohol use at grade 10 baseline wave 3}
\item{"can_0"}{cannabis use at grade 7 baseline wave 0, can; 0: No use; 1: used}
\item{"can_1"}{cannabis use at grade 8 baseline wave 1}
\item{"can_2"}{ annabis use at grade 9 baseline wave 2}
\item{"can_3"}{cannabis use at grade 10 baseline wave 3}
\item{"family_drug_use"}{family drug use at grade 7 baseline wave 0}
\item{"sex"}{sex; 0: male; 1: female}
\item{"illicit"}{Illicit drug use at age 25 (illicit: 0: No use; 1: used)}
}
}
\source{
\url{https://statsnotebook.io/blog/analysis/marginal_structural_model_iptw/}
}
\usage{
cannabis
}
\description{
Cannabis use at different time points - to be in marginal structural model with iptw to predict illict drug use at age 25
}
\keyword{datasets}
