% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.R
\docType{data}
\name{chic}
\alias{chic}
\title{Data of Chicago weather}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{city}{Chicago}
\item{date}{ddate}
\item{temp}{temperature}
\item{o3}{ozone}
\item{dewpoint}{dewpoint}
\item{pm10}{pm10}
\item{season}{season}
\item{yday}{day}
\item{month}{month}
\item{month_numeric}{numerical month}
\item{year}{year}
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
chic
}
\description{
Time series
}
\keyword{datasets}
