% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.704-package.R
\docType{package}
\name{data.704-package}
\alias{data.704}
\alias{data.704-package}
\title{data.704: Data for EBOH 704}
\description{
Supplies data for 704 labs and lectures
}
\author{
\strong{Maintainer}: Jay Brophy \email{james.brophy@mcgill.ca} (\href{https://orcid.org/0000-0001-8049-6875}{ORCID})

}
\keyword{internal}
