% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pow}
\alias{pow}
\title{Print x raised to the power y}
\usage{
pow(x, y)
}
\arguments{
\item{x, y}{Numbers.}
}
\value{
"invisible" \code{NULL}.
}
\description{
This function prints x raised to the power y for some reason (it's a test)
}
