% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summ}
\alias{summ}
\title{Print x + y}
\usage{
summ(x, y)
}
\arguments{
\item{x, y}{Numbers.}
}
\value{
"invisible" \code{NULL}.
}
\description{
This function prints x + y for some reason (it's a test)
}
