% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidyOR}
\alias{tidyOR}
\title{Tidyverse function to extract effect measures & 95\%CI}
\usage{
tidyOR(x)
}
\arguments{
\item{glm}{Model}
}
\value{
"invisible" \code{NULL}.
}
\description{
This function takes glm() as input or glm() \%>\% tidyOR
}
