% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/village_randomized.R
\docType{data}
\name{village_randomized}
\alias{village_randomized}
\title{A simulated randomized trial dataset}
\format{
A tibble with 1000 rows and 8 variables:
\describe{
\item{id}{Identification}
\item{sex}{Participant sex}
\item{age}{in years}
\item{pre_income}{Income before the intervention}
\item{program}{Randomized intervention}
\item{post_income}{Income after the intervention}
\item{sex_num}{Male = 1, female = 0}
\item{program_num}{0 = no promgram, 1 = Program}
}
}
\source{
\url{https://evalf20.classes.andrewheiss.com/example/rcts/}
}
\usage{
village_randomized
}
\description{
In this simulation, an NGO is planning on launching a training program designed to boost incomes.
Several potential confounders have been recorded
}
\keyword{datasets}
