% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_bern}
\alias{bin_bern}
\title{Create long data frame}
\usage{
bin_bern(df, Successes, Trials)
}
\arguments{
\item{df}{Dataframe}

\item{Successes, Trials}{Columns.}
}
\value{
"invisible" \code{NULL}.
}
\description{
This function takes data presented in Binomial (successes, trials) format and converts to Bernoulli (0,1) format
}
